Rename `inbox` module to `chan`
With the introduction of a `Mailbox`, having an `inbox` is confusing
when this module actually houses xtra's channel implementation.
