+++ src/send_future.rs
@@ -7,9 +7,9 @@ use std::task::{Context, Poll};
 use futures_core::FusedFuture;
 use futures_util::FutureExt;
 
+use crate::chan::{MailboxFull, MessageToAll, MessageToOne, RefCounter, WaitingSender};
 use crate::envelope::{BroadcastEnvelopeConcrete, ReturningEnvelope};
-use crate::inbox::{MailboxFull, MessageToAll, MessageToOne, RefCounter, WaitingSender};
-use crate::{inbox, Error, Handler};
+use crate::{chan, Error, Handler};
 
 /// A [`Future`] that represents the state of sending a message to an actor.
 ///
@@ -86,7 +86,7 @@ where
     /// Construct a [`SendFuture`] that contains the actor's name in its type.
     ///
     /// Compared to [`SendFuture::sending_erased`], this function avoids one allocation.
-    pub(crate) fn sending_named<M>(message: M, sender: inbox::ChanPtr<A, Rc>) -> Self
+    pub(crate) fn sending_named<M>(message: M, sender: chan::ChanPtr<A, Rc>) -> Self
     where
         A: Handler<M, Return = R>,
         M: Send + 'static,
@@ -104,7 +104,7 @@ where
 }
 
 impl<R> SendFuture<ActorErasedSending, ResolveToHandlerReturn<R>> {
-    pub(crate) fn sending_erased<A, M, Rc>(message: M, sender: inbox::ChanPtr<A, Rc>) -> Self
+    pub(crate) fn sending_erased<A, M, Rc>(message: M, sender: chan::ChanPtr<A, Rc>) -> Self
     where
         Rc: RefCounter,
         A: Handler<M, Return = R>,
@@ -127,7 +127,7 @@ impl<A, Rc> SendFuture<ActorNamedBroadcasting<A, Rc>, Broadcast>
 where
     Rc: RefCounter,
 {
-    pub(crate) fn broadcast_named<M>(msg: M, sender: inbox::ChanPtr<A, Rc>) -> Self
+    pub(crate) fn broadcast_named<M>(msg: M, sender: chan::ChanPtr<A, Rc>) -> Self
     where
         A: Handler<M, Return = ()>,
         M: Clone + Send + Sync + 'static,
@@ -146,7 +146,7 @@ where
 
 #[allow(dead_code)] // This will useful later.
 impl SendFuture<ActorErasedSending, Broadcast> {
-    pub(crate) fn broadcast_erased<A, M, Rc>(msg: M, sender: inbox::ChanPtr<A, Rc>) -> Self
+    pub(crate) fn broadcast_erased<A, M, Rc>(msg: M, sender: chan::ChanPtr<A, Rc>) -> Self
     where
         Rc: RefCounter,
         A: Handler<M, Return = ()>,
@@ -168,7 +168,7 @@ impl SendFuture<ActorErasedSending, Broadcast> {
 enum Sending<A, M, Rc: RefCounter> {
     New {
         msg: M,
-        sender: inbox::ChanPtr<A, Rc>,
+        sender: chan::ChanPtr<A, Rc>,
     },
     WaitingToSend(WaitingSender<M>),
     Done,
